Removed bounds check from Position
The Stage has to check if the Snake is out of bounds. It has to do so
anyway because besides checking whether the coordinates are >= 0 the
other <=-bound has to be checked, too.

When it is possible to create a Position with negative coordinates, the
snake can move to this prohibited position and the stage can check
whether it has gone off-stage.